Android balances hold numbers, not strings
AssetBalance.balance was Balance<String>, so every reader parsed it:
fifteen BigInteger() calls in the model alone, plus the swap, fiat,
asset detail and banner paths, and hasAvailable() swallowed a malformed
value as false. It is Balance<BigInteger> now, matching iOS, and the
text columns convert once at the persistence boundary. The asset info
query defaults the balance columns so a wallet with no balance row maps
without a null dance.

+++ android/data/services/store/src/main/kotlin/com/gemwallet/android/data/service/store/database/entities/DbBalance.kt
@@ -9,6 +9,7 @@ import com.gemwallet.android.ext.toAssetId
 import com.gemwallet.android.model.AssetBalance
 import com.gemwallet.android.model.Balance
 import com.wallet.core.primitives.BalanceMetadata
+import java.math.BigInteger
 
 @Entity(
     tableName = "balances",
@@ -82,16 +83,16 @@ fun DbBalance.toDTO(): AssetBalance? {
     return AssetBalance(
         asset = assetId.toAssetId()?.chain?.asset() ?: return null,
         balance = Balance(
-            available = available,
-            frozen = frozen,
-            locked = locked,
-            staked = staked,
-            pending = pending,
-            rewards = rewards,
-            reserved = reserved,
-            withdrawable = withdrawable,
-            pendingUnconfirmed = pendingUnconfirmed,
-            earn = earn,
+            available = available.toBigInteger(),
+            frozen = frozen.toBigInteger(),
+            locked = locked.toBigInteger(),
+            staked = staked.toBigInteger(),
+            pending = pending.toBigInteger(),
+            rewards = rewards.toBigInteger(),
+            reserved = reserved.toBigInteger(),
+            withdrawable = withdrawable.toBigInteger(),
+            pendingUnconfirmed = pendingUnconfirmed.toBigInteger(),
+            earn = earn.toBigInteger(),
         ),
         balanceAmount = Balance(
             available = availableAmount,
